Revert rename to nwpu
